makes changes to texture/materials in MDB and binary
materials are realized to act as a collection of various texture/colors and params. due to the decision to have the runtime database be different than the binary database the opportunity to export materials separate from their texture data is made possible as the runtime may now act as a liason between the two. it also allows for texture reuse without duplication of data inside of the binary files