continues to develop the table and runtime database structures
table now has begin and end range based iterators as well as specialized template iterators. the teamplte iterators allow unique typed Key Value tables to be search with fewer instructions as the compiler is able to deduce which field it is that we want at compile time. for equal types key value tables the default iterators may be used and a boolean is supplied to determine if a key or value search is being done. depending on the search type the appropriate structure member address is returned w/o branching