adds finer grainularity to simulation sync - now syncs only the states so that the simulation may be sync'd during multiple phases or defereed until later with try_write
change also makes the database a little more clear. The database will hold a simulation snapshot structure that holds the most up to date information about the simulation. the database will have various push and pull methods for the kinds of simulation data which may be pushed. will also be exploring this paradigm for the rendering side of things and terrain as well