moving back to messaging system while allocator design cooks on the back burner
implementing a very pendantic messaging system. The idea is Event Spaces are generalized names in which events can occur. events are triggred by actions. actions are predefined processes that may be performed during run time. when an event is triggered the action is perferformed and the event space will be notified of the relevant event type. any subscribers to the event space that are looking for the specific event filter(s) will process the message accordingly