deterrmined cause of phantom leaks
structure which dynamically allocate memory(std::string) where not being deallocated as their parent strutures (gex_metric) where never being deallocated but rather unwound on a ptr along with the allocating ddl_structure. will continue to go through and prune dynamically allocating structures. going to attempt to leave as much non transforming data as std::string as possible. things like the mappings and the file parsing may all be done once on a std::string but the parser may take that string memory and make any form of copy it needs to during the parsing process on the scratch pad