will still need rcu but looking at a "staging" double buffer
the database should be more active to justify its existence as a glorified manner. for this purpose i will be looking into an "update-in-parallel" model where the current stage is being prepared in an auxillary buffer of equal size to the currently visible buffer. a very bare bones generational rcu system will be used to allow concurrent access for reads while mainitaining the ability to syncrhonize the write and clean up(exchange of stage and commited). Database will alow only be used for mostly read only data to further justify its use case