more or less settling on original design
had a few philosophical hang ups on the design of the RCU and calling it and RCU. i believe many have been resolved execpt for an argument which must be made that the geneartionHandles[MAX_RCU_GENERATION] acts as the root pointer and the read-copy is on the "head" or old index of the generationHandles. am happy with design and performance. will be moving onto the database tables and finally hooking up the entire database system