passed message from Action to table pull
will continue to explore API further and work out certain work flow kinks. overall database design should be up and running. this test was only done on one thread but given the testing that both the messaging system and RCU have gone through i believe this is fine for now