physics node needed support for dynamic/static objects
issue remains with current implementation. will need ot think of a way to allow static objects to ignore the rigidbody. currently thinking to adding the dynamic flag to the collider, and if it is placing it into a tbale or mpa fr the rigid bdoies to look up and hook to when being read in off the file.