Adds proof of concept of mouse handling
left click is bound to the detect interaction function. when the context triggers the mouse callback the event is passed to the input system which routes the message to the mouse handler and performs the bound function.

the framework reserves the left and right mouse buttons. buttons 3-MAX_SUPPORT_BUTTONS is left for end user